pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "face-processing"
version = "0.2.0"
requires-python = ">=3.10"
dependencies = [
    "mediapipe>=0.10.9",
    "opencv-python>=4.8",
    "numpy>=1.24",
    "scipy>=1.11",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
log = ["pandas", "pyarrow"]
batch = ["pyyaml>=6.0"]
audio = ["librosa>=0.10", "scipy>=1.11", "numba"]

[project.scripts]
dataset-processing = "dataset_processing.cli:main"
call-video-preparation = "call_video_preparation.prepare:main"
call-video-cut = "call_video_preparation.cut:main"
call-video-restore = "call_video_preparation.restore:main"
face-batch-gdrive = "batch.gdrive_processor:main"
face-batch-local = "batch.local_processor:main"
face-batch-export = "batch.export_faceclip_batch:main"
